fix search functionality bug - not working after remounting searchbar component
