+++ client/src/components/MapButtons.jsx
@@ -1,13 +1,17 @@
 import React from 'react'
 
-const MapButtons = ({togglePolyLines, visibleTransit, showRoute, setShowRoute}) => {
+const MapButtons = ({togglePolyLines, visibleTransit, clearSearch, showRoute, setShowRoute}) => {
   return (
     <div>
 			<button onClick={togglePolyLines}>
 				{visibleTransit ? "Hide Transit Map" : "Show Transit Map"}
 			</button>
       
-			<button onClick={()=>setShowRoute(!showRoute)}>
+			<button onClick={()=>{
+				setShowRoute(!showRoute);
+				clearSearch()
+				}}
+			>
 				{showRoute ? "Cancel planning route" : "Plan Routes"}
 			</button>
